+++ webcurator-core/src/main/java/org/webcurator/core/coordinator/WctCoordinatorInterface.java
@@ -1,0 +1,7 @@
+package org.webcurator.core.coordinator;
+
+import org.webcurator.core.harvester.coordinator.HarvestCoordinator;
+import org.webcurator.core.store.coordinator.DigitalAssetStoreCoordinator;
+
+public interface WctCoordinator extends HarvestCoordinator, DigitalAssetStoreCoordinator {
+}
