+++ webcurator-core/src/main/java/org/webcurator/core/harvester/coordinator/HarvestLogManager.java
@@ -1,36 +1,212 @@
 package org.webcurator.core.harvester.coordinator;
 
 import java.io.File;
+import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.webcurator.core.exceptions.WCTRuntimeException;
+import org.webcurator.core.reader.LogReader;
+import org.webcurator.core.store.DigitalAssetStoreFactoryImpl;
 import org.webcurator.domain.model.core.LogFilePropertiesDTO;
 import org.webcurator.domain.model.core.TargetInstance;
 
-public interface HarvestLogManager {
+public class HarvestLogManagerImpl {
 
-	File getLogfile(TargetInstance aTargetInstance, String aFilename);
+	private HarvestAgentManagerImpl harvestAgentManager;
+	private DigitalAssetStoreFactoryImpl digitalAssetStoreFactory;
 
-	List<String> getHopPath(TargetInstance aTargetInstance, String aFileName, String aUrl);
+	private Logger log = LoggerFactory.getLogger(getClass());
 
-	List<String> getLogLinesByRegex(TargetInstance aTargetInstance, String aFileName, int aNoOfLines, String aRegex,
-			boolean prependLineNumbers);
+	public List<String> listLogFiles(TargetInstance targetInstance) {
+		checkNotNull(targetInstance);
 
-	Integer getFirstLogLineAfterTimeStamp(TargetInstance aTargetInstance, String aFileName, Long timestamp);
+		if (harvestAgentManager.runningOrPaused(targetInstance)) {
+			LogReader logReader = harvestAgentManager.getLogReader(targetInstance);
+			if (logReader == null) {
+				log.warn("list Log Files Failed. Failed to find the Log Reader for the Job {}.", targetInstance.getJobName());
+				return new ArrayList<String>();
+			}
+			return logReader.listLogFiles(targetInstance.getJobName());
+		} else {
+			// if not then check to see if the log files are available from the
+			// digital asset store.
+			LogReader logReader = digitalAssetStoreFactory.getLogReader();
+			return logReader.listLogFiles(targetInstance.getJobName());
+		}
+	}
 
-	Integer getFirstLogLineContaining(TargetInstance aTargetInstance, String aFileName, String match);
+	public List<LogFilePropertiesDTO> listLogFileAttributes(TargetInstance targetInstance) {
+		checkNotNull(targetInstance);
 
-	Integer getFirstLogLineBeginning(TargetInstance aTargetInstance, String aFileName, String match);
+		LogReader logReader = getLogReader(targetInstance);
+		if (logReader == null) {
+			log.warn("listLogFileAttributes Failed. Failed to find the Log Reader for the Job {}.", targetInstance.getJobName());
+			return new ArrayList<>();
+		}
+		return logReader.listLogFileAttributes(targetInstance.getJobName());
+	}
 
-	List<String> getLog(TargetInstance aTargetInstance, String aFileName, int aStartLine, int aNoOfLines);
+	public List<String> tailLog(TargetInstance targetInstance, String aFileName, int aNoOfLines) {
+		checkNotNull(targetInstance);
 
-	Integer countLogLines(TargetInstance aTargetInstance, String aFileName);
+		LogReader logReader = getLogReader(targetInstance);
+		if (logReader == null) {
+			log.warn("Tail Log Files Failed. Failed to find the log Reader for the Job {}.", targetInstance.getJobName());
+			return Collections.singletonList("");
+		}
+		return logReader.tail(targetInstance.getJobName(), aFileName, aNoOfLines);
+	}
 
-	List<String> headLog(TargetInstance aTargetInstance, String aFileName, int aNoOfLines);
+	public Integer countLogLines(TargetInstance targetInstance, String aFileName) {
+		checkNotNull(targetInstance);
 
-	List<LogFilePropertiesDTO> listLogFileAttributes(TargetInstance aTargetInstance);
+		Integer count = 0;
+		LogReader logReader = getLogReader(targetInstance);
+		if (logReader == null) {
+			log.warn("Count Log Lines Failed. Failed to find the log Reader for the Job {}.", targetInstance.getJobName());
+			return count;
+		}
+		return logReader.countLines(targetInstance.getJobName(), aFileName);
+	}
 
-	List<String> listLogFiles(TargetInstance aTargetInstance);
+	public List<String> headLog(TargetInstance targetInstance, String aFileName, int aNoOfLines) {
+		checkNotNull(targetInstance);
 
-	List<String> tailLog(TargetInstance targetInstance, String aFileName, int aNoOfLines);
+		LogReader logReader = getLogReader(targetInstance);
+		if (logReader == null) {
+			log.warn("Head Log Files Failed. Failed to find the log Reader for the Job {}.", targetInstance.getJobName());
+			return Collections.singletonList("");
+		}
+		return logReader.get(targetInstance.getJobName(), aFileName, 1, aNoOfLines);
+	}
 
+
+	public List<String> getLog(TargetInstance targetInstance, String aFileName, int aStartLine, int aNoOfLines) {
+		checkNotNull(targetInstance);
+
+		LogReader logReader = getLogReader(targetInstance);
+		if (logReader == null) {
+			log.warn("Get Log Files Failed. Failed to find the log Reader for the Job {}.", targetInstance.getJobName());
+			return Collections.singletonList("");
+		}
+		return logReader.get(targetInstance.getJobName(), aFileName, aStartLine, aNoOfLines);
+	}
+
+
+	public Integer getFirstLogLineBeginning(TargetInstance targetInstance, String aFileName, String match) {
+		checkNotNull(targetInstance);
+
+		LogReader logReader = getLogReader(targetInstance);
+		if (logReader == null) {
+			log.warn("Get First Log Line Beginning failed. Failed to find the log Reader for the Job {}.",
+					targetInstance.getJobName());
+			return 0;
+		}
+		return logReader.findFirstLineBeginning(targetInstance.getJobName(), aFileName, match);
+	}
+
+	public Integer getFirstLogLineContaining(TargetInstance targetInstance, String aFileName, String match) {
+		checkNotNull(targetInstance);
+
+		LogReader logReader = getLogReader(targetInstance);
+		if (logReader == null) {
+			log.warn("Get First Log Line Containing failed. Failed to find the log Reader for the Job {}.",
+					targetInstance.getJobName());
+			return 0;
+		}
+		return logReader.findFirstLineContaining(targetInstance.getJobName(), aFileName, match);
+	}
+
+
+	public Integer getFirstLogLineAfterTimeStamp(TargetInstance targetInstance, String aFileName, Long timestamp) {
+		checkNotNull(targetInstance);
+
+		LogReader logReader = getLogReader(targetInstance);
+		if (logReader == null) {
+			log.warn("Get First Log Line After Timestamp failed. Failed to find the log Reader for the Job {}.",
+					targetInstance.getJobName());
+			return 0;
+		}
+		return logReader.findFirstLineAfterTimeStamp(targetInstance.getJobName(), aFileName, timestamp);
+	}
+
+
+	public List<String> getLogLinesByRegex(TargetInstance targetInstance, String aFileName, int aNoOfLines, String aRegex,
+			boolean prependLineNumbers) {
+		checkNotNull(targetInstance);
+
+		LogReader logReader = getLogReader(targetInstance);
+		if (logReader == null) {
+			log.warn("Get log lines by regex failed. Failed to find the log Reader for the Job {}.", targetInstance.getJobName());
+			return Collections.singletonList("");
+		}
+		return logReader.getByRegularExpression(targetInstance.getJobName(), aFileName, aRegex, "zzzzzzzzz", prependLineNumbers, 0,
+				aNoOfLines);
+	}
+
+
+	public List<String> getHopPath(TargetInstance targetInstance, String aFileName, String aUrl) {
+		checkNotNull(targetInstance);
+
+		LogReader logReader = getLogReader(targetInstance);
+		if (logReader == null) {
+			log.warn("Get Log Files Failed. Failed to find the log Reader for the Job {}.", targetInstance.getJobName());
+			return Collections.singletonList("");
+		}
+		return logReader.getHopPath(targetInstance.getJobName(), targetInstance.getHarvestResult(1).getOid().toString(), aFileName,
+				aUrl);
+	}
+
+
+	public File getLogfile(TargetInstance targetInstance, String aFilename) {
+		checkNotNull(targetInstance);
+
+		LogReader logReader = getLogReader(targetInstance);
+		if (logReader == null) {
+			log.warn("Tail Log Files Failed. Failed to find the log Reader for the Job {}.", targetInstance.getJobName());
+			return null;
+		}
+		return logReader.retrieveLogfile(targetInstance.getJobName(), aFilename);
+	}
+
+	private void checkNotNull(TargetInstance targetInstance) {
+		if (targetInstance == null) {
+			throw new WCTRuntimeException("Target instance must not be null");
+		}
+	}
+
+	private LogReader getLogReader(TargetInstance targetInstance) {
+		if (harvestAgentManager.runningOrPaused(targetInstance)) {
+			// If we are harvesting then get the log files from the harvester
+			LogReader logReader = harvestAgentManager.getLogReader(targetInstance);
+			return logReader;
+		} else {
+			// if not then check to see if the log files are available from the
+			// digital asset store.
+			LogReader logReader = digitalAssetStoreFactory.getLogReader();
+			return logReader;
+		}
+
+	}
+
+	public HarvestAgentManagerImpl getHarvestAgentManager() {
+		return harvestAgentManager;
+	}
+
+	public void setHarvestAgentManagerImpl(HarvestAgentManagerImpl harvestAgentManager) {
+		this.harvestAgentManager = harvestAgentManager;
+	}
+
+	public DigitalAssetStoreFactoryImpl getDigitalAssetStoreFactory() {
+		return digitalAssetStoreFactory;
+	}
+
+   public void setDigitalAssetStoreFactory(DigitalAssetStoreFactoryImpl digitalAssetStoreFactory) {
+		this.digitalAssetStoreFactory = digitalAssetStoreFactory;
+	}
+	
+	
 }
